Refactored Catering with RecommendationWizard;
Fixes and tweaks;
Added more typescript related fixes;
