+++ package.json
@@ -19,6 +19,7 @@
     "@types/react-datepicker": "^3.1.2",
     "@types/react-dom": "^17.0.8",
     "@types/react-redux": "^7.1.16",
+    "@types/react-router-dom": "^5.1.7",
     "date-fns": "^2.16.1",
     "date-fns-tz": "^1.0.12",
     "env-cmd": "^10.1.0",
@@ -43,7 +44,7 @@
     "react-spinners": "^0.9.0",
     "react-swipeable": "^6.1.0",
     "react-transition-group": "^4.4.1",
-    "typescript": "^4.3.4"
+    "typescript": "^4.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
